--- src/main/java/com/dnd/namuiwiki/application/oauth/OAuthProviderService.java
@@ -1,7 +1,0 @@
-package com.dnd.namuiwiki.application.oauth;
-
-import com.dnd.namuiwiki.application.oauth.dto.OAuthUserInfoDto;
-
-public interface OAuthProviderService {
-    OAuthUserInfoDto getOAuthUserInfo(String code);
-}
