--- src/main/java/com/dnd/namuiwiki/application/oauth/kakao/dto/KakaoOAuthUserInfoResponse.java
@@ -1,20 +1,0 @@
-package com.dnd.namuiwiki.application.oauth.kakao.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@ToString
-@AllArgsConstructor
-public class KakaoOAuthUserInfoResponse {
-
-    private Long id;
-    private Boolean has_signed_up;
-    private String connected_at;
-    private String synched_at;
-    private String properties;
-    private Object kakao_account;
-    private Object for_partner;
-
-}
