--- src/main/java/com/dnd/namuiwiki/application/oauth/type/OAuthProvider.java
@@ -1,19 +1,0 @@
-package com.dnd.namuiwiki.application.oauth.type;
-
-import lombok.Getter;
-
-@Getter
-public enum OAuthProvider {
-    KAKAO("KAKAO");
-
-    private String name;
-
-    private OAuthProvider(String name) {
-        this.name = name;
-    }
-
-    public static OAuthProvider of(String provider) {
-        return OAuthProvider.valueOf(provider.toUpperCase());
-    }
-
-}
