+++ src/main/java/com/dnd/namuiwiki/domain/jwt/dto/TokenUserInfoDto.java
@@ -1,0 +1,10 @@
+package com.dnd.namuiwiki.domain.jwt.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class TokenUserInfoDto {
+    private String wikiId;
+}
