+++ src/main/java/com/dnd/namuiwiki/domain/oauth/OAuthProviderService.java
@@ -1,0 +1,7 @@
+package com.dnd.namuiwiki.domain.oauth;
+
+import com.dnd.namuiwiki.domain.oauth.dto.OAuthUserInfoDto;
+
+public interface OAuthProviderService {
+    OAuthUserInfoDto getOAuthUserInfo(String code);
+}
