+++ src/main/java/com/dnd/namuiwiki/domain/oauth/kakao/dto/KakaoOAuthTokenResponse.java
@@ -1,0 +1,20 @@
+package com.dnd.namuiwiki.domain.oauth.kakao.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@ToString
+@AllArgsConstructor
+public class KakaoOAuthTokenResponse {
+
+    private String token_type;
+    private String access_token;
+    private String id_token;
+    private Long expires_in;
+    private String refresh_token;
+    private Long refresh_token_expires_in;
+    private String scope;
+
+}
