--- src/main/java/com/dnd/namuiwiki/jwt/TokenUserInfoDto.java
@@ -1,10 +1,0 @@
-package com.dnd.namuiwiki.jwt;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class TokenUserInfoDto {
-    private String wikiId;
-}
