--- src/main/java/com/dnd/namuiwiki/presentation/auth/dto/OAuthLoginResponse.java
@@ -1,13 +1,0 @@
-package com.dnd.namuiwiki.presentation.auth.dto;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class OAuthLoginResponse {
-
-    private final String accessToken;
-    private final String refreshToken;
-
-}
