--- src/main/java/com/dnd/namuiwiki/user/User.java
@@ -1,29 +1,0 @@
-package com.dnd.namuiwiki.user;
-
-import com.dnd.namuiwiki.application.oauth.type.OAuthProvider;
-import lombok.Getter;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-import java.util.UUID;
-
-@Getter
-@Document(collection = "users")
-public class User {
-    @Id
-    private String id;
-    private String wikiId;
-    private OAuthProvider oAuthProvider;
-    private String oAuthId;
-    private String refreshToken;
-
-    public User(OAuthProvider oAuthProvider, String oAuthId) {
-        this.wikiId = UUID.randomUUID().toString();
-        this.oAuthProvider = oAuthProvider;
-        this.oAuthId = oAuthId;
-    }
-
-    public void setRefreshToken(String refreshToken) {
-        this.refreshToken = refreshToken;
-    }
-}
