--- src/main/java/com/dnd/namuiwiki/user/UserRepository.java
@@ -1,10 +1,0 @@
-package com.dnd.namuiwiki.user;
-
-import com.dnd.namuiwiki.application.oauth.type.OAuthProvider;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends MongoRepository<User, String> {
-    Optional<User> findByOAuthProviderAndOAuthId(OAuthProvider oAuthProvider, String oAuthId);
-}
