+++ src/test/java/com/dnd/namuiwiki/jwt/JwtProviderTest.java
@@ -1,5 +1,7 @@
 package com.dnd.namuiwiki.jwt;
 
+import com.dnd.namuiwiki.domain.jwt.JwtProvider;
+import com.dnd.namuiwiki.domain.jwt.dto.TokenUserInfoDto;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
 import org.junit.jupiter.api.Test;
